requirements.txt
fastapi
python-dotenv
google-genai
uvicorn